+++ packages/atproto/test/src/services/identity_service_test.dart
@@ -8,6 +8,7 @@ import 'package:atproto_core/atproto_core.dart' as core;
 // 🌎 Project imports:
 import 'package:atproto/src/ids.g.dart';
 import 'package:atproto/src/services/entities/did.dart';
+import 'package:atproto/src/services/entities/did_credentials.dart';
 import 'package:atproto/src/services/entities/plc_operation.dart';
 import 'suite/service_suite.dart';
 
@@ -30,5 +31,15 @@ void main() {
   testIdentity<PlcOperation>(
     (m, s) => s.signPlcOperation(),
     id: comAtprotoIdentitySignPlcOperation,
+  );
+
+  testIdentity<DidCredentials>(
+    (m, s) => s.getRecommendedDidCredentials(),
+    id: comAtprotoIdentityGetRecommendedDidCredentials,
+  );
+
+  testIdentity<core.EmptyData>(
+    (m, s) => s.requestPlcOperationSignature(),
+    id: comAtprotoIdentityRequestPlcOperationSignature,
   );
 }